+++ pylate/evaluation/__init__.py
@@ -2,8 +2,9 @@ from .beir import evaluate, get_beir_triples, load_beir
 from .colbert_distillation import ColBERTDistillationEvaluator
 from .colbert_triplet import ColBERTTripletEvaluator
 from .custom_dataset import load_custom_dataset
-from .InformationRetrievalEvaluator import PyLateInformationRetrievalEvaluator
-from .nanoBEIR import NanoBEIREvaluator
+from .nano_beir_evaluator import NanoBEIREvaluator
+from .pylate_information_retrieval_evaluator import \
+    PyLateInformationRetrievalEvaluator
 
 __all__ = [
     "ColBERTTripletEvaluator",
