+++ pylate/evaluation/nano_beir_evaluator.py
